+++ sam/data_sources/__init__.py
@@ -1,6 +1,6 @@
 from .synthetic_data import synthetic_timeseries, synthetic_date_range
-from .knmy_stations import knmy_stations
-from .weather import read_knmi, read_openweathermap, read_regenradar
+from .weather import read_knmi, read_openweathermap, \
+                     read_regenradar, read_knmi_station_data
 from .mongo_wrapper import MongoWrapper
 from .deprecated import create_synthetic_timeseries
 
@@ -9,4 +9,5 @@ from . import synthetic_data
 from . import weather
 
 __all__ = ["synthetic_timeseries", "synthetic_date_range", "read_knmi",
-           "read_openweathermap", "read_regenradar", "MongoWrapper"]
+           "read_knmi_station_data", "read_openweathermap", "read_regenradar",
+           "MongoWrapper"]