+++ CHANGELOG.md
@@ -6,6 +6,13 @@ From version 2.1.0 on, versioning scheme should be as follows: Version x.y.z
 - y for new features
 - z for bugfixes
 
+## Version 2.1.1
+
+### Changes
+- `sam.data_sources.read_knmi` was changed because of a new KNMI API. The package `knmy` does not work anymore. 
+- `sam.data_sources.read_knmi_station_data` was added to get KNMI data for a selection of KNMI station numbers
+- `knmy` is no longer a (optional) dependency
+
 ## Version 2.1.0
 
 ### New features